Version 4.2
Added a pleasant (i hope not too bad at least) selection of background (shades) along with the one that was there before. Now, on hour 1 the back- ground sticks to a normal (looking from dark-light) blue to light-yellow scheme, and on hour 2 as usual in reverse (same but going from light to dark), then on hour 3 going to light again the background shades green to light-blue, then in reverse on hour 4, then finally on 5 going to light the background shades green to light-pink, then in reverse in reverse on hour 6.