Version 5.3
Fine-tuned the visibility protection of Comment so that it now looks great around the dark background area. Previously was looking a bit muggy. Stretched this protection exactly around the perceptional boundary area and fined-tuned the 'back-darkening' effect at this lower range as well so that it now just does that job it should do.