Version 6.3
Removed the jarring black vs. white effect that the Normal identifier had, now this simply brightens a bit when the background and it start to clash. It now looks reasonably pleasing to the eye as time passes, rather than abruptly change from black->white and vice versa as it did previously.