Version 6.7
Made a fix so that the background now lightens and darkens, lightens and darkens, etc.. using a six-color rotation  different pastel shades: light-yellow, light-green, light-red, light-blue, light cyan, and light magenta. This means that the background now lightens-up using a certain shade, darkens down in another shade, then lightens-up in yet another, and so on. for a total of six different colours over 3 hours. I've tried to make it so that the background simulates the way the real sky colours at morning time and then colours at evening time and then again at morning time, and so on.