Version 8.6
Simplified the logic to just 'did i move to cause a timer previously, and if so where, hence move the opposite way'. Now it's pretty flawless apart from when you're on the final line and this line has no characters and the size of the file is only 3 lines long. This not only causes a flash/beep and the timer not to work (hence no automatic colour changes) but also when you try moving off this line it could go funny (the cursor mysteriously jumps up a line) but it's nothing serious.