Version 16.5
Revved up the visibility of PreProc. PreProc is the 'function' keyword in php so it now brightens nicer and doesn't become hard-to-see. Similarly with Identifier, and Special.  Identifier in PHP obviously means the variables. Not sure if its used in other languages but at least if you use PHP it now looks nice and strong.